Change size() to return size_t
There are a lot of uses of size() that quickly start using integer math
(losing precision) but size() is okay in C++ as long as we’re willing
to use exceptions.  If we’re not, we can revisit after we get
everything compiling.

We should start converting the failed conversions into int with some
comment (or util function) so we can easily find and revert all of
these cases if/when we change our mind.

+++ org/antlr/v4/runtime/UnbufferedCharStream.cpp
@@ -189,7 +189,7 @@ namespace org {
                     }
                 }
 
-                int UnbufferedCharStream::size() { 
+                size_t UnbufferedCharStream::size() {
                     throw UnsupportedOperationException(std::wstring(L"Unbuffered stream cannot know its size"));
                 }
